backend/requirements.txt
llama-parse>=0.4.0
python-dotenv>=1.0.0
python-docx>=1.0.0
pypdf>=3.17.0
mongoengine>=0.27.0
pymongo>=4.6.0
loguru
google-genai

openai>=1.12.0
aiolimiter>=1.1.0
backoff>=2.2.1
tiktoken>=0.5.2
langchain-experimental>=0.0.50
langchain-text-splitters>=0.0.1
langchain_google_genai>=2.1.8
langchain_openai>=0.3.28
langchain-core>=0.1.20
pinecone>=3.0.0

fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0

# Dependencies for multilingual-e5-large model
# transformers>=4.30.0
# torch>=2.0.0
# sentence-transformers>=2.2.2 #didn't
